pyproject.toml
[project]
name = "result-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=2.6.27",
    "jupyter>=1.1.1",
    "marker-pdf>=1.8.2",
    "mistralai>=1.9.3",
    "pymupdf>=1.24.1",
]
